spider-web/pyproject.toml
[tool.poetry]
name = "spider-web"
version = "0.1.0"
description = ""
authors = ["Ronald Perdigão <git.ronaldseabra@gmail.com>"]
readme = "README.md"
packages = [{include = "spider_web"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.21.1"
psycopg2 = "^2.9.5"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
ipython = "^8.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | .docker/volumes
  | project-support
)/
'''

[tool.isort]
profile = "black"
